chore: delete tracethickness01.fixture.tsx and trace-thickness.test.ts
