+++ examples/features/trace-thickness/trace-thickness-01.fixture.tsx
