+++ tests/trace-thickness3.test.ts
@@ -1,0 +1,16 @@
+import { test, expect } from "bun:test"
+import {
+  getTraceThicknessFromConnection,
+  STANDARD_TRACE_THICKNESS,
+} from "../lib/utils/getTraceThicknessFromConnection"
+
+test("getTraceThicknessFromConnection - default", () => {
+  const connection = {
+    name: "test",
+    pointsToConnect: [],
+  }
+
+  expect(getTraceThicknessFromConnection(connection)).toBe(
+    STANDARD_TRACE_THICKNESS,
+  )
+})
