+++ tests/trace-thickness5.test.ts
@@ -1,0 +1,12 @@
+import { test, expect } from "bun:test"
+import { validateTraceThicknessParameters } from "../lib/utils/getTraceThicknessFromConnection"
+
+test("validateTraceThicknessParameters - valid", () => {
+  const connection = {
+    name: "test",
+    pointsToConnect: [],
+    traceThickness: 0.3,
+  }
+
+  expect(validateTraceThicknessParameters(connection)).toEqual([])
+})
